+++ packages/gamut/CHANGELOG.md
@@ -3,6 +3,16 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [67.6.0](https://github.com/Codecademy/gamut/compare/@codecademy/gamut@67.5.4...@codecademy/gamut@67.6.0) (2025-12-10)
+
+### Features
+
+- **ConnectedForm, GridForm:** Add aria-controls to nested checkboxes ([2a7a7ad](https://github.com/Codecademy/gamut/commit/2a7a7ade029972c2e7f990cc140d270b17123d9f))
+
+### [67.5.4](https://github.com/Codecademy/gamut/compare/@codecademy/gamut@67.5.3...@codecademy/gamut@67.5.4) (2025-12-08)
+
+**Note:** Version bump only for package @codecademy/gamut
+
 ### [67.5.3](https://github.com/Codecademy/gamut/compare/@codecademy/gamut@67.5.2...@codecademy/gamut@67.5.3) (2025-12-05)
 
 **Note:** Version bump only for package @codecademy/gamut