+++ packages/styleguide/CHANGELOG.md
@@ -3,6 +3,10 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+### [78.5.2](https://github.com/Codecademy/gamut/compare/@codecademy/styleguide@78.5.1...@codecademy/styleguide@78.5.2) (2025-12-08)
+
+**Note:** Version bump only for package @codecademy/styleguide
+
 ### [78.5.1](https://github.com/Codecademy/gamut/compare/@codecademy/styleguide@78.5.0...@codecademy/styleguide@78.5.1) (2025-12-04)
 
 ### Bug Fixes